requirement.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11
facebook-sdk==2.0.0
httplib2==0.11.0
idna==2.6
oauth2==1.9.0.post1
oauthlib==2.0.7
PySocks==1.6.8
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
